refactor(metadata): move to `media::metadata`
we no longer model a `Metadata` ty directly in the database! as such, it's weird to have a `model` import for it.
