+++ src/models/media/builder/image_crate.rs
@@ -9,7 +9,10 @@ use sqlx::types::Json;
 
 use crate::{
     error::RavesError,
-    models::metadata::{builder::get_video_len, Format, MediaKind, SpecificMetadata},
+    models::media::{
+        builder::get_video_len,
+        metadata::{Format, MediaKind, SpecificMetadata},
+    },
 };
 
 use super::MediaBuilder;
