+++ src/models/media/builder/kamadak.rs
@@ -5,9 +5,9 @@ use tokio::try_join;
 
 use crate::{
     error::RavesError,
-    models::metadata::{
-        builder::get_video_len, Format, MediaKind, OtherMetadataMap, OtherMetadataValue,
-        SpecificMetadata,
+    models::media::{
+        builder::get_video_len,
+        metadata::{Format, MediaKind, OtherMetadataMap, OtherMetadataValue, SpecificMetadata},
     },
 };
