+++ src/models/media/builder/matroska.rs
@@ -5,7 +5,7 @@
 
 use crate::{
     error::RavesError,
-    models::metadata::{Format, MediaKind, SpecificMetadata},
+    models::media::metadata::{Format, MediaKind, SpecificMetadata},
 };
 
 use camino::Utf8Path;
