+++ src/models/media/builder/nom.rs
@@ -6,9 +6,9 @@ use tokio::try_join;
 
 use crate::{
     error::RavesError,
-    models::metadata::{
-        builder::get_video_len, Format, MediaKind, OtherMetadataMap, OtherMetadataValue,
-        SpecificMetadata,
+    models::media::{
+        builder::get_video_len,
+        metadata::{Format, MediaKind, OtherMetadataMap, OtherMetadataValue, SpecificMetadata},
     },
 };
