+++ src/models/mod.rs
@@ -1,4 +1,3 @@
 pub mod media;
-pub mod metadata;
 pub mod tags;
 pub mod thumbnail;
